Add cast to all XCTAssertEqual for Travis CI.

+++ Project/CollectionUtilsTests/CUCompactArrayTests.m
@@ -49,7 +49,7 @@
     NSArray *array = [self sampleArray];
     NSArray *compactArray = [array cu_compactArray];
     
-    XCTAssertEqual(compactArray.count, 4);
+    XCTAssertEqual(compactArray.count, (NSUInteger)4);
 }
 
 - (void)testNestedArray
@@ -66,8 +66,8 @@
                        @"4"];
     NSArray *compactArray = [array cu_compactArray];
     
-    XCTAssertEqual(compactArray.count, 5);
-    XCTAssertEqual([compactArray[3] count], 3);
+    XCTAssertEqual(compactArray.count, (NSUInteger)5);
+    XCTAssertEqual([compactArray[3] count], (NSUInteger)3);
 }
 
 #pragma mark - primitive instance methods
@@ -77,7 +77,7 @@
     NSArray *array = [self sampleArray];
     NSArray *compactArray = [array cu_compactArray];
     
-    XCTAssertEqual(compactArray.count, 4);
+    XCTAssertEqual(compactArray.count, (NSUInteger)4);
 }
 
 - (void)testObjectAtIndex
@@ -178,18 +178,18 @@
     NSArray *array = [self sampleArray];
     NSArray *compactArray = [array cu_compactArray];
     
-    XCTAssertEqual([compactArray indexOfObject:@"0"], 0);
-    XCTAssertEqual([compactArray indexOfObject:@"1"], 1);
-    XCTAssertEqual([compactArray indexOfObject:@"2"], 2);
-    XCTAssertEqual([compactArray indexOfObject:@"3"], 3);
-    XCTAssertEqual([compactArray indexOfObject:@"4"], NSNotFound);
-    XCTAssertEqual([compactArray indexOfObject:[NSNull null]], NSNotFound);
-    XCTAssertEqual([compactArray indexOfObject:nil], NSNotFound);
+    XCTAssertEqual([compactArray indexOfObject:@"0"], (NSUInteger)0);
+    XCTAssertEqual([compactArray indexOfObject:@"1"], (NSUInteger)1);
+    XCTAssertEqual([compactArray indexOfObject:@"2"], (NSUInteger)2);
+    XCTAssertEqual([compactArray indexOfObject:@"3"], (NSUInteger)3);
+    XCTAssertEqual([compactArray indexOfObject:@"4"], (NSUInteger)NSNotFound);
+    XCTAssertEqual([compactArray indexOfObject:[NSNull null]], (NSUInteger)NSNotFound);
+    XCTAssertEqual([compactArray indexOfObject:nil], (NSUInteger)NSNotFound);
     
-    XCTAssertEqual([compactArray indexOfObject:@"0" inRange:NSMakeRange(0, 3)], 0);
-    XCTAssertEqual([compactArray indexOfObject:@"1" inRange:NSMakeRange(0, 3)], 1);
-    XCTAssertEqual([compactArray indexOfObject:@"2" inRange:NSMakeRange(0, 3)], 2);
-    XCTAssertEqual([compactArray indexOfObject:@"3" inRange:NSMakeRange(0, 3)], NSNotFound);
+    XCTAssertEqual([compactArray indexOfObject:@"0" inRange:NSMakeRange(0, 3)], (NSUInteger)0);
+    XCTAssertEqual([compactArray indexOfObject:@"1" inRange:NSMakeRange(0, 3)], (NSUInteger)1);
+    XCTAssertEqual([compactArray indexOfObject:@"2" inRange:NSMakeRange(0, 3)], (NSUInteger)2);
+    XCTAssertEqual([compactArray indexOfObject:@"3" inRange:NSMakeRange(0, 3)], (NSUInteger)NSNotFound);
 }
 
 - (void)testIndexOfObjectIdenticalTo
@@ -197,16 +197,16 @@
     NSArray *array = [self sampleArray];
     NSArray *compactArray = [array cu_compactArray];
     
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"0"], 0);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"1"], 1);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"2"], 2);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"3"], 3);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"4"], NSNotFound);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"0"], (NSUInteger)0);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"1"], (NSUInteger)1);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"2"], (NSUInteger)2);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"3"], (NSUInteger)3);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"4"], (NSUInteger)NSNotFound);
     
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"0" inRange:NSMakeRange(0, 3)], 0);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"1" inRange:NSMakeRange(0, 3)], 1);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"2" inRange:NSMakeRange(0, 3)], 2);
-    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"3" inRange:NSMakeRange(0, 3)], NSNotFound);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"0" inRange:NSMakeRange(0, 3)], (NSUInteger)0);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"1" inRange:NSMakeRange(0, 3)], (NSUInteger)1);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"2" inRange:NSMakeRange(0, 3)], (NSUInteger)2);
+    XCTAssertEqual([compactArray indexOfObjectIdenticalTo:@"3" inRange:NSMakeRange(0, 3)], (NSUInteger)NSNotFound);
 }
 
 - (void)testIsEqualToArray
@@ -468,14 +468,14 @@ static NSComparisonResult stringSort(id str1, id str2, void *context)
         return [str isEqualToString:@"2"];
     }];
     
-    XCTAssertEqual(index, 2);
+    XCTAssertEqual(index, (NSUInteger)2);
     
     index = [compactArray indexOfObjectWithOptions:NSEnumerationConcurrent | NSEnumerationReverse passingTest:^BOOL(id obj, NSUInteger idx, BOOL *stop) {
         NSString *str = obj;
         return [str isEqualToString:@"2"];
     }];
     
-    XCTAssertEqual(index, 2);
+    XCTAssertEqual(index, (NSUInteger)2);
     
     NSMutableIndexSet *indexSet = [NSMutableIndexSet indexSet];
     [indexSet addIndex:0];
@@ -487,7 +487,7 @@ static NSComparisonResult stringSort(id str1, id str2, void *context)
         return [str isEqualToString:@"2"];
     }];
     
-    XCTAssertEqual(index, 2);
+    XCTAssertEqual(index, (NSUInteger)2);
 }
 
 - (void)testIndexesOfObjectsPassingTest
@@ -579,7 +579,7 @@ static NSComparisonResult stringSort(id str1, id str2, void *context)
         return [str1 compare:str2];
     }];
     
-    XCTAssertEqual(index, 3);
+    XCTAssertEqual(index, (NSUInteger)3);
 }
 
 #pragma mark - NSCopying
